pyproject.toml
[project]
name = "neuralstyletransfer-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
]
